chore 🔥: move model to nn.cv
