+++ fedhf/model/nn/cv/alexnet_cifar10.py
@@ -9,7 +9,7 @@
 import torch
 import torch.nn as nn
 
-from .base_model import BaseModel
+from ..base_model import BaseModel
 
 
 class AlexNetCIFAR10(BaseModel):
