+++ fedhf/model/nn/cv/densenet.py
@@ -10,7 +10,7 @@ import torch
 import torch.nn as nn
 from torchvision import models
 
-from .base_model import BaseModel
+from ..base_model import BaseModel
 
 
 class DenseNet(BaseModel):
